backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
bcrypt
pandas
pdfplumber
openpyxl
google-genai
openai
pydantic-settings
jinja2
